fix(server/v2): respect context cancellation in start command
When running external tests that involve starting a v2 server, the only
way to get the command to shut down was through sending an interrupt or
termination signal, which is not possible to do independently through
many concurrent serer commands within the same process.

Respect the root context being cancelled, as a signal that the server
needs to be stopped.

Unfortunately there are no existing unit tests in the server/v2 package
with running a start command, that can be expanded to cover this
behavior.