refactor(simdv2): allow non-comet server components
This renames NewRootCmd to NewCometBFTRootCmd, to make it clear that it
uses CometBFT for the consensus layer. Callers who want to use a
different consensus-providing server component can instead call
NewRootCmdWithConsensusComponent, passing in a callback to be evaluated
with a client context in order to produce the component.

This pattern is working for Gordian-Cosmos integration.

I'm not especially tied to the new function names. I'm also open to more
clear naming for "consensus component" which occurs a few times.